spaceship tunes module require optimizations/fixes
