+++ spaceship/lib/spaceship/tunes/app_details.rb
@@ -1,3 +1,7 @@
+require_relative 'errors'
+require_relative 'language_item'
+require_relative 'tunes_base'
+
 module Spaceship
   module Tunes
     class AppDetails < TunesBase
@@ -77,7 +81,7 @@ module Spaceship
       # Push all changes that were made back to iTunes Connect
       def save!
         client.update_app_details!(application.apple_id, raw_data)
-      rescue Spaceship::TunesClient::ITunesConnectError => ex
+      rescue Spaceship::Tunes::Error => ex
         if ex.to_s == "operation_failed"
           # That's alright, we get this error message if nothing has changed
         else
