+++ spaceship/lib/spaceship/tunes/app_image.rb
@@ -1,3 +1,5 @@
+require_relative 'tunes_base'
+
 module Spaceship
   module Tunes
     # Represents an image hosted on iTunes Connect. Used for icons, screenshots, etc
