+++ spaceship/lib/spaceship/tunes/app_submission.rb
@@ -1,3 +1,4 @@
+require_relative 'tunes_base'
 
 module Spaceship
   module Tunes
