+++ spaceship/lib/spaceship/tunes/app_version_history.rb
@@ -1,3 +1,5 @@
+require_relative 'app_version_states_history'
+
 module Spaceship
   module Tunes
     # Represents a read only version of an iTunes Connect Versions State History
