+++ spaceship/lib/spaceship/tunes/app_version_ref.rb
@@ -1,3 +1,5 @@
+require_relative 'tunes_base'
+
 module Spaceship
   module Tunes
     class AppVersionRef < TunesBase
