+++ spaceship/lib/spaceship/tunes/build_details.rb
@@ -1,3 +1,5 @@
+require_relative 'tunes_base'
+
 module Spaceship
   module Tunes
     # Represents the details of a build
