+++ spaceship/lib/spaceship/tunes/recovery_device.rb
@@ -1,3 +1,5 @@
+require_relative 'tunes_base'
+
 module Spaceship
   module Tunes
     class RecoveryDevice < TunesBase
