+++ spaceship/lib/spaceship/tunes/sandbox_tester.rb
@@ -1,3 +1,5 @@
+require_relative 'tunes_base'
+
 module Spaceship
   module Tunes
     class SandboxTester < TunesBase
