+++ spaceship/lib/spaceship/tunes/territory.rb
@@ -1,3 +1,5 @@
+require_relative 'tunes_base'
+
 module Spaceship
   module Tunes
     class Territory < TunesBase
