+++ spaceship/lib/spaceship/tunes/transit_app_file.rb
@@ -1,3 +1,5 @@
+require_relative 'tunes_base'
+
 module Spaceship
   module Tunes
     # Represents a geo json
