+++ spaceship/lib/spaceship/tunes/tunes.rb
@@ -1,40 +1,35 @@
-require 'spaceship/tunes/tunes_base'
-require 'spaceship/tunes/application'
-require 'spaceship/tunes/iap'
-require 'spaceship/tunes/version_set'
-require 'spaceship/tunes/app_version'
-require 'spaceship/tunes/app_version_common'
-require 'spaceship/tunes/app_submission'
-require 'spaceship/tunes/tunes_client'
-require 'spaceship/tunes/language_item'
-require 'spaceship/tunes/app_status'
-require 'spaceship/tunes/app_image'
-require 'spaceship/tunes/app_ratings'
-require 'spaceship/tunes/app_version_ref'
-require 'spaceship/tunes/app_version_history'
-require 'spaceship/tunes/app_version_states_history'
-require 'spaceship/tunes/transit_app_file'
-require 'spaceship/tunes/user_detail'
-require 'spaceship/tunes/app_screenshot'
-require 'spaceship/tunes/language_converter'
-require 'spaceship/tunes/build'
-require 'spaceship/tunes/build_details'
-require 'spaceship/tunes/build_train'
-require 'spaceship/tunes/device_type'
-require 'spaceship/tunes/app_trailer'
-require 'spaceship/tunes/sandbox_tester'
-require 'spaceship/tunes/app_details'
-require 'spaceship/tunes/pricing_tier'
-require 'spaceship/tunes/recovery_device'
-require 'spaceship/tunes/territory'
-require 'spaceship/tunes/availability'
-require 'spaceship/tunes/members'
-require 'spaceship/tunes/member'
+require_relative 'tunes_base'
+require_relative 'application'
+require_relative 'iap'
+require_relative 'version_set'
+require_relative 'app_version'
+require_relative 'app_version_common'
+require_relative 'app_submission'
+require_relative 'tunes_client'
+require_relative 'language_item'
+require_relative 'app_status'
+require_relative 'app_image'
+require_relative 'app_ratings'
+require_relative 'app_version_ref'
+require_relative 'app_version_history'
+require_relative 'app_version_states_history'
+require_relative 'transit_app_file'
+require_relative 'user_detail'
+require_relative 'app_screenshot'
+require_relative 'language_converter'
+require_relative 'build'
+require_relative 'build_details'
+require_relative 'build_train'
+require_relative 'device_type'
+require_relative 'app_trailer'
+require_relative 'sandbox_tester'
+require_relative 'app_details'
+require_relative 'pricing_tier'
+require_relative 'recovery_device'
+require_relative 'territory'
+require_relative 'availability'
+require_relative 'members'
+require_relative 'member'
 
-require 'spaceship/tunes/app_version_promocodes'
-require 'spaceship/tunes/app_version_generated_promocodes'
-
-# File Uploads
-require 'spaceship/du/utilities'
-require 'spaceship/du/upload_file'
-require 'spaceship/du/du_client'
+require_relative 'app_version_promocodes'
+require_relative 'app_version_generated_promocodes'
