+++ spaceship/lib/spaceship/tunes/user_detail.rb
@@ -1,3 +1,5 @@
+require_relative 'tunes_base'
+
 module Spaceship
   module Tunes
     class UserDetail < TunesBase
